Move export_optimized_onnx_model to backend.py
